Set Curated to true on CVE 2010-1870
Appears to be a dead end vulnerability from a previous semester. A previous student filled out some of the information, but left the rest blank and curated to false. I set curated to true, per Professor Meneely's request.

I'm considering this to be a dead end and I'll look into the other two CVE's assigned to me.